@@ -1,15 +1,14 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-	<groupId>de.muenchen</groupId>
 	<artifactId>mailmerge</artifactId>
-	<version>0.0.1-SNAPSHOT</version>
 	<name>Seriendruck</name>
-
-	<properties>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<libreoffice.version>4.1</libreoffice.version>
-	</properties>
+	
+	<parent>
+		<groupId>de.muenchen</groupId>
+		<artifactId>mailmerge-parent</artifactId>
+		<version>0.0.1-SNAPSHOT</version>
+	</parent>
 
 	<build>
 		<resources>
@@ -23,9 +22,9 @@
 				<artifactId>maven-compiler-plugin</artifactId>
 				<version>3.5.1</version>
 				<configuration>
-					<compilerVersion>1.7</compilerVersion>
-					<source>1.7</source>
-					<target>1.7</target>
+					<compilerVersion>1.8</compilerVersion>
+					<source>1.8</source>
+					<target>1.8</target>
 					<encoding>UTF-8</encoding>
 				</configuration>
 			</plugin>
@@ -172,7 +171,7 @@
 		<dependency>
 			<groupId>de.muenchen</groupId>
 			<artifactId>wollmux-core</artifactId>
-			<version>16.11-SNAPSHOT</version>
+			<version>18.1.0-SNAPSHOT</version>
 		</dependency>
 		<dependency>
 			<groupId>de.muenchen</groupId>
