@@ -8,12 +8,17 @@ import javax.swing.JFrame;
 import javax.swing.JLabel;
 import javax.swing.SwingUtilities;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import de.muenchen.allg.itd51.wollmux.core.util.L;
-import de.muenchen.allg.itd51.wollmux.core.util.Logger;
 import de.muenchen.mailmerge.dialog.Common;
 
 public class MailMergeProgressWindow
 {
+
+  private static final Logger LOGGER = LoggerFactory.getLogger(MailMergeProgressWindow.class);
+
   private JFrame myFrame;
 
   private JLabel countLabel;
@@ -27,39 +32,33 @@ public class MailMergeProgressWindow
     this.maxcount = maxcount;
     try
     {
-      SwingUtilities.invokeAndWait(new Runnable()
-      {
-        @Override
-        public void run()
-        {
-          myFrame = new JFrame(L.m("Seriendruck"));
-          Common.setWollMuxIcon(myFrame);
-          Box vbox = Box.createVerticalBox();
-          myFrame.getContentPane().add(vbox);
-          Box hbox = Box.createHorizontalBox();
-          vbox.add(hbox);
-          hbox.add(Box.createHorizontalStrut(5));
-          hbox.add(new JLabel(L.m("Verarbeite Dokument")));
-          hbox.add(Box.createHorizontalStrut(5));
-          countLabel = new JLabel("   -");
-          hbox.add(countLabel);
-          hbox.add(new JLabel(" / " + maxcount + "    "));
-          hbox.add(Box.createHorizontalStrut(5));
-          myFrame.setAlwaysOnTop(true);
-          myFrame.pack();
-          int frameWidth = myFrame.getWidth();
-          int frameHeight = myFrame.getHeight();
-          Dimension screenSize = Toolkit.getDefaultToolkit().getScreenSize();
-          int x = screenSize.width / 2 - frameWidth / 2;
-          int y = screenSize.height / 2 - frameHeight / 2;
-          myFrame.setLocation(x, y);
-          myFrame.setVisible(true);
-        }
+      SwingUtilities.invokeAndWait(() -> {
+        myFrame = new JFrame(L.m("Seriendruck"));
+        Common.setWollMuxIcon(myFrame);
+        Box vbox = Box.createVerticalBox();
+        myFrame.getContentPane().add(vbox);
+        Box hbox = Box.createHorizontalBox();
+        vbox.add(hbox);
+        hbox.add(Box.createHorizontalStrut(5));
+        hbox.add(new JLabel(L.m("Verarbeite Dokument")));
+        hbox.add(Box.createHorizontalStrut(5));
+        countLabel = new JLabel("   -");
+        hbox.add(countLabel);
+        hbox.add(new JLabel(" / " + maxcount + "    "));
+        hbox.add(Box.createHorizontalStrut(5));
+        myFrame.setAlwaysOnTop(true);
+        myFrame.pack();
+        int frameWidth = myFrame.getWidth();
+        int frameHeight = myFrame.getHeight();
+        Dimension screenSize = Toolkit.getDefaultToolkit().getScreenSize();
+        int x = screenSize.width / 2 - frameWidth / 2;
+        int y = screenSize.height / 2 - frameHeight / 2;
+        myFrame.setLocation(x, y);
+        myFrame.setVisible(true);
       });
-    }
-    catch (Exception x)
+    } catch (Exception x)
     {
-      Logger.error(x);
+      LOGGER.error("", x);
     }
   }
 
@@ -67,22 +66,15 @@ public class MailMergeProgressWindow
   {
     try
     {
-      SwingUtilities.invokeLater(new Runnable()
-      {
-        @Override
-        public void run()
-        {
-          ++count;
-          countLabel.setText("" + count);
-          if (maxcount > 0)
-            myFrame.setTitle("" + Math.round(100 * (double) count / maxcount)
-              + "%");
-        }
+      SwingUtilities.invokeLater(() -> {
+        ++count;
+        countLabel.setText("" + count);
+        if (maxcount > 0)
+          myFrame.setTitle("" + Math.round(100 * (double) count / maxcount) + "%");
       });
-    }
-    catch (Exception x)
+    } catch (Exception x)
     {
-      Logger.error(x);
+      LOGGER.error("", x);
     }
   }
 
@@ -90,18 +82,10 @@ public class MailMergeProgressWindow
   {
     try
     {
-      SwingUtilities.invokeLater(new Runnable()
-      {
-        @Override
-        public void run()
-        {
-          myFrame.dispose();
-        }
-      });
-    }
-    catch (Exception x)
+      SwingUtilities.invokeLater(() -> myFrame.dispose());
+    } catch (Exception x)
     {
-      Logger.error(x);
+      LOGGER.error("", x);
     }
   }
 }
