@@ -30,12 +30,14 @@
  */
 package de.muenchen.mailmerge.print;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import de.muenchen.allg.itd51.wollmux.XPrintModel;
 import de.muenchen.allg.itd51.wollmux.core.functions.ExternalFunction;
 import de.muenchen.allg.itd51.wollmux.core.parser.ConfigThingy;
 import de.muenchen.allg.itd51.wollmux.core.parser.ConfigurationErrorException;
-import de.muenchen.allg.itd51.wollmux.core.util.Logger;
-import de.muenchen.mailmerge.WollMuxClassLoader;
+import de.muenchen.mailmerge.MailMergeClassLoader;
 
 /**
  * Eine durch ein ConfigThingy beschriebene externe Druckfunktion.
@@ -45,6 +47,8 @@ import de.muenchen.mailmerge.WollMuxClassLoader;
 public class PrintFunction implements Comparable<PrintFunction>
 {
 
+  private static final Logger LOGGER = LoggerFactory.getLogger(PrintFunction.class);
+
   private ExternalFunction func = null;
 
   private String functionName;
@@ -53,18 +57,18 @@ public class PrintFunction implements Comparable<PrintFunction>
 
   /**
    * Erzeugt aus einem ConfigThingy (übergeben wird der EXTERN-Knoten) eine
-   * PrintFunction vom Namen functionName mit dem ORDER-Wert order. Die Werte order
-   * und functionName werden für korrekte Sortierung verschiedener
-   * Druckfuntionktionen und damit für die Bildung einer definierten Reihenfolge bei
-   * der Abarbeitung verschachtelter Druckfunktionen benätigt (siehe compareTo(...)).
+   * PrintFunction vom Namen functionName mit dem ORDER-Wert order. Die Werte
+   * order und functionName werden für korrekte Sortierung verschiedener
+   * Druckfuntionktionen und damit für die Bildung einer definierten Reihenfolge
+   * bei der Abarbeitung verschachtelter Druckfunktionen benätigt (siehe
+   * compareTo(...)).
    * 
    * @throws ConfigurationErrorException
    *           falls die Spezifikation in conf fehlerhaft ist.
    */
-  public PrintFunction(ConfigThingy conf, String functionName, int order)
-      throws ConfigurationErrorException
+  public PrintFunction(ConfigThingy conf, String functionName, int order) throws ConfigurationErrorException
   {
-    func = new ExternalFunction(conf, WollMuxClassLoader.getClassLoader());
+    func = new ExternalFunction(conf, MailMergeClassLoader.getClassLoader());
     this.functionName = functionName;
     this.order = order;
   }
@@ -84,8 +88,9 @@ public class PrintFunction implements Comparable<PrintFunction>
    * einem eigenen Thread) auf und liefert den Thread zurück.
    * 
    * @param pmod
-   *          das XPrintModel des aktuellen Vordergrunddokuments, das die wichtigsten
-   *          Druckkomandos bereitstellt, die die externe Funktion verwenden kann.
+   *          das XPrintModel des aktuellen Vordergrunddokuments, das die
+   *          wichtigsten Druckkomandos bereitstellt, die die externe Funktion
+   *          verwenden kann.
    * @throws Exception
    */
   public Thread invoke(XPrintModel pmod)
@@ -93,18 +98,13 @@ public class PrintFunction implements Comparable<PrintFunction>
     final Object[] args;
     args = new Object[] { pmod };
 
-    Thread t = new Thread(new Runnable()
-    {
-      public void run()
+    Thread t = new Thread(() -> {
+      try
       {
-        try
-        {
-          func.invoke(args);
-        }
-        catch (java.lang.Exception e)
-        {
-          Logger.error(e);
-        }
+        func.invoke(args);
+      } catch (java.lang.Exception e)
+      {
+        LOGGER.error("", e);
       }
     });
     t.start();
@@ -112,14 +112,15 @@ public class PrintFunction implements Comparable<PrintFunction>
   }
 
   /**
-   * Vergleicht this mit otherPrintFunction und liefert -1 zurück, wenn this eine
-   * höhere Priorität besitzt als otherPrintFunction (und damit vor
-   * otherPrintFunction abgearbeitet werden soll) und 1, wenn otherPrintFunction eine
-   * höhere Priorität besitzt als this. Die Priorität ergibt sich dabei aus dem
-   * Attribut ORDER der PrintFunction und deren Namen. Ist die this.order kleiner als
-   * otherPrintFunction.order, so hat this die höhrer Priorität, sind beide
-   * order-Werte gleich, so wird der Name alphabetisch verglichen und ist this.order
-   * größer als otherPrintFunction.order, so hat otherPrintFunction hörere Priorität.
+   * Vergleicht this mit otherPrintFunction und liefert -1 zurück, wenn this
+   * eine höhere Priorität besitzt als otherPrintFunction (und damit vor
+   * otherPrintFunction abgearbeitet werden soll) und 1, wenn otherPrintFunction
+   * eine höhere Priorität besitzt als this. Die Priorität ergibt sich dabei aus
+   * dem Attribut ORDER der PrintFunction und deren Namen. Ist die this.order
+   * kleiner als otherPrintFunction.order, so hat this die höhrer Priorität,
+   * sind beide order-Werte gleich, so wird der Name alphabetisch verglichen und
+   * ist this.order größer als otherPrintFunction.order, so hat
+   * otherPrintFunction hörere Priorität.
    * 
    * @param otherPrintFunction
    *          Die PrintFunction mit der vergleichen werden soll.
@@ -127,25 +128,30 @@ public class PrintFunction implements Comparable<PrintFunction>
    * 
    * @author Christoph Lutz (D-III-ITD-5.1)
    */
+  @Override
   public int compareTo(PrintFunction otherPrintFunction)
   {
     PrintFunction other = otherPrintFunction;
-    if (this.order != other.order) return (this.order < other.order) ? -1 : 1;
+    if (this.order != other.order)
+      return (this.order < other.order) ? -1 : 1;
     return this.functionName.compareTo(other.functionName);
   }
 
+  @Override
   public boolean equals(Object o)
   {
-    if (o == null) return false;
+    if (o == null)
+      return false;
     try
     {
       return (this.compareTo((PrintFunction) o) == 0);
+    } catch (ClassCastException x)
+    {
     }
-    catch (ClassCastException x)
-    {}
     return false;
   }
 
+  @Override
   public int hashCode()
   {
     return this.functionName.hashCode();
@@ -156,6 +162,7 @@ public class PrintFunction implements Comparable<PrintFunction>
    * 
    * @see java.lang.Object#toString()
    */
+  @Override
   public String toString()
   {
     return "PrintFunction['" + functionName + "']";
