@@ -32,6 +32,9 @@
 import java.net.URL;
 import java.net.URLClassLoader;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import com.sun.star.drawing.XDrawPageSupplier;
 import com.sun.star.text.XTextDocument;
 import com.sun.star.text.XTextSectionsSupplier;
@@ -40,7 +43,6 @@ import com.sun.star.uno.UnoRuntime;
 
 import de.muenchen.allg.afid.UNO;
 import de.muenchen.allg.itd51.wollmux.core.util.L;
-import de.muenchen.allg.itd51.wollmux.core.util.Logger;
 import de.muenchen.allg.itd51.wollmux.core.util.Utils;
 
 /**
@@ -52,6 +54,9 @@ import de.muenchen.allg.itd51.wollmux.core.util.Utils;
  */
 public class Workarounds
 {
+
+  private static final Logger LOGGER = LoggerFactory.getLogger(Workarounds.class);
+
   private static Boolean workaround89783 = null;
 
   private static Boolean workaround102164 = null;
@@ -64,7 +69,7 @@ public class Workarounds
 
   public static Boolean applyWorkaround(String issueNumber)
   {
-    Logger.debug("Workaround für Issue "
+    LOGGER.debug("Workaround für Issue "
       + issueNumber
       + " aktiv. Bestimmte Features sind evtl. nicht verfügbar. Die Performance kann ebenfalls leiden.");
     return Boolean.TRUE;
@@ -165,8 +170,11 @@ public class Workarounds
   {
     if (workaround89783 == null)
     {
-      Logger.debug(L.m("Workaround für TDF Issue 89783 aktiv."));
+      LOGGER.debug(L.m("Workaround für TDF Issue 89783 aktiv."));
       workaround89783 = true;
+    } else
+    {
+      workaround89783 = false;
     }
 
     if(workaround89783)
