@@ -38,6 +38,9 @@ package de.muenchen.mailmerge.event;
 import java.util.HashSet;
 import java.util.Set;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import com.sun.star.beans.PropertyValue;
 import com.sun.star.frame.DispatchDescriptor;
 import com.sun.star.frame.FrameAction;
@@ -52,7 +55,6 @@ import com.sun.star.util.URL;
 
 import de.muenchen.allg.afid.UNO;
 import de.muenchen.allg.itd51.wollmux.core.util.L;
-import de.muenchen.allg.itd51.wollmux.core.util.Logger;
 
 /**
  * Liefert zu Dispatch-URLs, die der WollMux behandeln kann XDispatch-Objekte.
@@ -62,6 +64,9 @@ import de.muenchen.allg.itd51.wollmux.core.util.Logger;
 public class DispatchProviderAndInterceptor implements XDispatchProvider,
     XDispatchProviderInterceptor
 {
+
+  private static final Logger LOGGER = LoggerFactory.getLogger(DispatchProviderAndInterceptor.class);
+
   /**
    * Enthält einen XDispatchProvider, der Dispatch-Objekte für alle globalen (d.h.
    * nicht dokumentgebundenen) Funktionalitäten des WollMux bereitstellt.
@@ -243,7 +248,7 @@ public class DispatchProviderAndInterceptor implements XDispatchProvider,
     {
       DispatchProviderAndInterceptor dpi = new DispatchProviderAndInterceptor(frame);
 
-      Logger.debug(L.m("Registriere DocumentDispatchInterceptor #%1 für frame #%2",
+      LOGGER.debug(L.m("Registriere DocumentDispatchInterceptor #%1 für frame #%2",
         Integer.valueOf(dpi.hashCode()), Integer.valueOf(frame.hashCode())));
 
       UNO.XDispatchProviderInterception(frame).registerDispatchProviderInterceptor(
@@ -254,7 +259,7 @@ public class DispatchProviderAndInterceptor implements XDispatchProvider,
       frame.addFrameActionListener(dpi.frameActionListener);
     }
     else
-      Logger.debug(L.m(
+      LOGGER.debug(L.m(
         "Ignoriere doppelten Aufruf von registerDocumentDispatchInterceptor() für den selben Frame #%1",
         Integer.valueOf(frame.hashCode())));
   }
@@ -292,7 +297,7 @@ public class DispatchProviderAndInterceptor implements XDispatchProvider,
             dpi.frameActionListener = null;
           }
 
-          Logger.debug(L.m(
+          LOGGER.debug(L.m(
             "Deregistrierung von DocumentDispatchInterceptor #%1 aus frame #%2",
             Integer.valueOf(dpi.hashCode()), Integer.valueOf(dpi.frame.hashCode())));
           UNO.XDispatchProviderInterception(dpi.frame).releaseDispatchProviderInterceptor(
@@ -334,7 +339,7 @@ public class DispatchProviderAndInterceptor implements XDispatchProvider,
       return;
     synchronized (documentDispatchProviderAndInterceptors)
     {
-      Logger.debug(L.m("Interne Freigabe des DocumentDispatchInterceptor #%1",
+      LOGGER.debug(L.m("Interne Freigabe des DocumentDispatchInterceptor #%1",
         Integer.valueOf(dpi.hashCode())));
       documentDispatchProviderAndInterceptors.remove(dpi);
     }
